Separated video functions into own .c file.
